requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.11
django-filter==24.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.10.0
uritemplate==4.1.1
